Attempts to stop services
Implemented a method for attempting to stop services.

Updated a few user prompts so that they understand the actions they are about to take.

Attempting to implement a method to terminate processes but am unsuccessful so far.

Removed unregistering DLL files.  After some research the only thing registering DLL files does is to create certain registry entries for that said DLL.  Which will be removed anyway by the regshot diff log so it is pointless.

Reorganized the code and have the -F switch deleting registry values before registry keys so that it will throwing errors (less errors if there's any).